+++ pywatershed/hydrology/PRMSCanopy.py
@@ -3,7 +3,7 @@ from numba import prange
 
 from ..base.adapter import adaptable
 from ..base.control import Control
-from ..base.storageUnit import StorageUnit
+from ..base.storage_unit import StorageUnit
 from ..constants import CovType, HruType, numba_num_threads, zero
 from ..parameters import Parameters
