+++ pywatershed/hydrology/PRMSEt.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pywatershed.base.budget import Budget
-from pywatershed.base.storageUnit import StorageUnit
+from pywatershed.base.storage_unit import StorageUnit
 
 from ..base.adapter import adaptable
 from ..base.control import Control
