+++ pywatershed/hydrology/PRMSGroundwater.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from pywatershed.base.storageUnit import StorageUnit
+from pywatershed.base.storage_unit import StorageUnit
 
 from ..base.adapter import adaptable
 from ..base.control import Control
