+++ pywatershed/hydrology/PRMSSoilzone.py
@@ -1,7 +1,7 @@
 import numpy as np
 from numba import prange
 
-from pywatershed.base.storageUnit import StorageUnit
+from pywatershed.base.storage_unit import StorageUnit
 
 from ..base.adapter import adaptable
 from ..base.control import Control
